+++ package.json
@@ -43,7 +43,8 @@
     "format": "./bin/prettier.js --write",
     "format:single": "npm run format -- src/printer.js",
     "format:all": "npm run format -- index.js src/*.js bin/*.js",
-    "build:docs": "rollup -c docs/rollup.config.js"
+    "build:docs": "rollup -c docs/rollup.config.js",
+    "build:dist": "rollup -c dist/rollup.config.js"
   },
   "jest": {
     "setupFiles": [
